Split `InstanceLabelSelectClauseFactory.ts` into multiple source files by implementation
